# [symex] assume assertions after checking to strengthen path conditions (EXPERIMENTAL!!!)