--- a/regression/esbmc-unix2/20_flex/test.desc
+++ b/regression/esbmc-unix2/20_flex/test.desc
@@ -1,4 +1,4 @@
-KNOWNBUG
+CORE
 main.c
 --unwind 1 --ir --z3
 ^VERIFICATION FAILED$