This PR adds `assume(new_expr)` after `assertion(new_expr, msg)` to strengthen path conditions with proven properties. This appears to facilitate improved SMT solver constraint propagation and path pruning. The approach seems particularly effective for k-induction, where each proven assertion serves as an assumption for subsequent verification steps, enabling the SMT solver to eliminate infeasible paths earlier in the search. In particular:
- Base case: Proven assertions become assumptions for the next step
- Inductive step: Each assertion immediately strengthens the induction hypothesis
- Solver can eliminate infeasible paths much faster

Performance improvements on k-induction regression test suites:
- K-induction: 18.5% faster (2344→1909 sec*proc)
- Total regression: 12.3% faster (3823→3353 sec*proc)
- Real time: 10.7% faster (956→853 sec)