@lucasccordeiro 2025-07-25
Master (30s):

````
Statistics:          33569 Files
  correct:           17620
    correct true:    10341
    correct false:    7279
  incorrect:            34
    incorrect true:     13
    incorrect false:    21
  unknown:           15915
  Score:             27209 (max: 55885)

https://github.com/esbmc/esbmc/actions/runs/16404623531
````

This PR (30s):
````
Statistics:          33569 Files
  correct:           17473
    correct true:    10212
    correct false:    7261
  incorrect:            35
    incorrect true:     13
    incorrect false:    22
  unknown:           16061
  Score:             26917 (max: 55885)

https://github.com/esbmc/esbmc/actions/runs/16518833748/
````